Add files via upload
The online event notice board project centers on developing a digital platform that streamlines event management processes. This platform enables organizers to effortlessly announce event details, schedules, and updates of the events which was held on other colleges. It is benefit for those who are not attend the classes regularly, they can get updated every details of the events in this website . The emphasis lies in creating a user-friendly space for effective online event coordination and enhanced participant experiences.